+++ app/electron-client/src/fileAssociations.ts
@@ -6,30 +6,17 @@
  * process, and launching new instances of the IDE when necessary. The module also exports
  * constants related to file associations and project handling.
  */
+import type { Electron } from '@/electron'
+import type { Event } from 'electron'
+import * as common from 'enso-common/src/constants'
 import * as fsSync from 'node:fs'
 import * as pathModule from 'node:path'
-import process from 'node:process'
-
-import * as electron from 'electron'
-import electronIsDev from 'electron-is-dev'
-
-import * as common from 'enso-common'
-
 import * as project from 'project-manager-shim'
-
 import * as fileAssociations from '../fileAssociations'
 export * from '../fileAssociations'
 
-// =================
-// === Constants ===
-// =================
-
 /** Returned by {@link String.indexOf} when the substring was not found. */
 const NOT_FOUND = -1
-
-// ==========================
-// === Arguments Handling ===
-// ==========================
 
 /**
  * Check if the given list of application startup arguments denotes an attempt to open a file.
@@ -40,7 +27,7 @@ const NOT_FOUND = -1
  * executable name and any electron dev mode arguments.
  * @returns The path to the file to open, or `null` if no file was specified.
  */
-export function argsDenoteFileOpenAttempt(clientArgs: readonly string[]): string | null {
+export function getFileToOpen(clientArgs: readonly string[]): string | null {
   const arg = clientArgs[0]
   let result: string | null = null
   // If the application is invoked with exactly one argument and this argument is a file, we
@@ -57,11 +44,11 @@ export function argsDenoteFileOpenAttempt(clientArgs: readonly string[]): string
   return result
 }
 
-/** Get the arguments, excluding the initial program name and any electron dev mode arguments. */
-export const CLIENT_ARGUMENTS = getClientArguments()
-
-/** Decide what are client arguments, @see {@link CLIENT_ARGUMENTS}. */
-function getClientArguments(args = process.argv): readonly string[] {
+/** Parse client arguments. */
+export function parseClientArguments(
+  args: readonly string[],
+  electronIsDev: boolean,
+): readonly string[] {
   if (electronIsDev) {
     // Client arguments are separated from the electron dev mode arguments by a '--' argument.
     const separator = '--'
@@ -84,10 +71,6 @@ function getClientArguments(args = process.argv): readonly string[] {
   }
 }
 
-// =========================
-// === File Associations ===
-// =========================
-
 /** Check if the given path looks like a file that we can open. */
 export function isFileOpenable(path: string): boolean {
   const extension = pathModule.extname(path).toLowerCase()
@@ -98,7 +81,7 @@ export function isFileOpenable(path: string): boolean {
 }
 
 /** Callback called when a file is opened via the `open-file` event. */
-export function onFileOpened(event: electron.Event, path: string): string | null {
+export function onFileOpened(event: Event, path: string): string | null {
   console.log(`Received 'open-file' event for path '${path}'.`)
   if (isFileOpenable(path)) {
     console.log(`The file '${path}' is openable.`)
@@ -116,7 +99,10 @@ export function onFileOpened(event: electron.Event, path: string): string | null
  * if this IDE instance should load the project. See {@link onFileOpened} for more details.
  * @param setProjectToOpen - A function that will be called with the path of the project to open.
  */
-export function setOpenFileEventHandler(setProjectToOpen: (path: string) => void) {
+export function setOpenFileEventHandler(
+  setProjectToOpen: (path: string) => void,
+  electron: Electron,
+) {
   electron.app.on('open-file', (_event, path) => {
     console.log(`Opening file '${path}'.`)
     setProjectToOpen(path)
@@ -154,7 +140,7 @@ export function setOpenFileEventHandler(setProjectToOpen: (path: string) => void
  * @returns The ID of the project to open.
  * @throws {Error} if the project from the file cannot be opened or imported.
  */
-export function handleOpenFile(openedFile: string): project.ProjectInfo {
+export function handleOpenFile(openedFile: string, electron: Electron): project.ProjectInfo {
   try {
     const title = openedFile
       .split(pathModule.sep)