+++ CHANGELOG.md
@@ -28,6 +28,9 @@
   top.
 - [Context menu for connections][14325].
 - [Warnings and Errors no longer become transparent][14388]
+- [`--headless` flag to run a project without the User Interface][14310]
+- [Preview for assets (text, main files of Projects, audio, video, images) in
+  Right Sidebar][14310]
 - ["Delete and Connect Around" option in node's menu][14403]
 - [GeoMap visualization is now working without need of Mapbox Token in
   environment][14429]
@@ -57,6 +60,7 @@
 [14209]: https://github.com/enso-org/enso/pull/14209
 [14215]: https://github.com/enso-org/enso/pull/14215
 [14270]: https://github.com/enso-org/enso/pull/14270
+[14310]: https://github.com/enso-org/enso/pull/14310
 [14311]: https://github.com/enso-org/enso/pull/14311
 [14267]: https://github.com/enso-org/enso/pull/14267
 [14325]: https://github.com/enso-org/enso/pull/14325